Rename challenges/types.ts
