+++ src/challenges/c1-number-reader/index.ts
@@ -1,4 +1,4 @@
-import type { Challenge } from '..';
+import type { Challenge } from '../types';
 import solution from './solution';
 
 const input = 877209536;
