+++ src/challenges/c3-number-to-words/index.ts
@@ -1,4 +1,4 @@
-import { Challenge } from '..';
+import { Challenge } from '../types';
 import solution from './solution';
 
 const input = 743243892;
