+++ src/challenges/c4-roman-numerals/index.ts
@@ -1,4 +1,4 @@
-import { Challenge } from '..';
+import { Challenge } from '../types';
 import solution from './solution';
 
 const input = 2476;
