+++ src/challenges/c5-add/index.ts
@@ -1,4 +1,4 @@
-import { Challenge } from '..';
+import { Challenge } from '../types';
 import solution from './solution';
 
 const input: [number, number] = [64722, 87549];
