+++ src/challenges/c6-binary-tree/index.ts
@@ -1,4 +1,4 @@
-import { Challenge } from '..';
+import { Challenge } from '../types';
 import solution from './solution';
 
 interface Node {
